+++ packages/plugin-file-upload/web/src/icons/getIcon.tsx
@@ -75,7 +75,8 @@ function getIconFromList(type, checkList, typePredicate) {
 
 export const getIcon = type => {
   const icon =
-    getIconFromList(type, iconRegList, (regExp, type) => regExp.test(type)) ||
-    getIconFromList(type, iconList, (typeList, type) => typeList.some(e => e === type));
+    type &&
+    (getIconFromList(type, iconRegList, (regExp, type) => regExp.test(type)) ||
+      getIconFromList(type, iconList, (typeList, type) => typeList.some(e => e === type)));
   return icon || otherIcon;
 };