+++ CHANGELOG.md
@@ -19,8 +19,10 @@
 ### :bug: Bug Fix
 - `editor`
   - [#1463](https://github.com/wix-incubator/rich-content/pull/1463) external formatting toolbar: preserve selection on inline style change
+  - fixtate 'draftjs-conductor' version due to broken es5
 - `file-upload`
   - [#1465](https://github.com/wix-incubator/rich-content/pull/1465) remove dependency of file type, long extension clipping and otherIcon themeing fix
+
 
 </details>
 <hr/>